requirements.txt
Flask
google-cloud-firestore
firebase-admin
google-generativeai
grpcio
google-cloud-core